refactor(prf): keep the PRF crate backend-agnostic
The crate now ships only the abstraction: traits, context and encoding
domains, visitor machinery, and ReadyPrf, with no cryptographic
dependencies. The local HMAC-SHA256 backend and its integration tests
land separately as the vitaminc-hmac crate (#276) so the implementation
can be reviewed and hardened on its own terms.

Unit tests run against a deterministic non-cryptographic mock backend
that implements the full Prf contract, proving the trait contract
stands alone. Also registers vitaminc-prf in the release-plz version
group, which it had been missing from since the crate was introduced.
