+++ packages/prf/README.md
@@ -6,78 +6,18 @@ domain-separated pseudorandom values. Inputs cross backend boundaries in
 backend output is always awaitable so local and remote batched implementations
 share one interface.
 
-The local HMAC-SHA256 backend delegates HMAC construction to RustCrypto and
-keeps the keyed state behind `ProtectedDigest`.
+This crate defines only the abstraction: the [`PrfValue`](https://docs.rs/vitaminc-prf/latest/vitaminc_prf/trait.PrfValue.html)
+and [`Prf`](https://docs.rs/vitaminc-prf/latest/vitaminc_prf/trait.Prf.html)
+traits, context and encoding domains, and the visitor machinery. It contains
+no cryptography. Concrete backends live in their own crates; the local
+HMAC-SHA256 backend is [`vitaminc-hmac`](https://docs.rs/vitaminc-hmac),
+and its documentation carries runnable end-to-end examples.
 
 Every protected leaf carries an explicit [`PrfEncoding`](https://docs.rs/vitaminc-prf/latest/vitaminc_prf/struct.PrfEncoding.html)
 domain. Built-in text, bytes, and fixed-width integers are separated even when
 their byte representations happen to match. Custom leaf implementations must
 provide a stable, namespaced encoding identifier, preventing accidental
 untagged derivation.
-
-## Deriving a block
-
-Built-in values return the backend's raw block through `prf`. Use
-`prf_with_context` to separate the same value across indexes or applications.
-
-```rust
-use vitaminc_prf::{HmacSha256Prf, PrfValue};
-use vitaminc_protected::Protected;
-
-# async fn example() -> Result<(), Box<dyn std::error::Error>> {
-let prf = HmacSha256Prf::new(Protected::new(vec![7; 32]));
-let term = "alice@example.com"
-    .prf_with_context(prf, "users/email/exact/v1")
-    .await?;
-assert_eq!(term.len(), 32);
-# Ok(())
-# }
-```
-
-## Producing application-specific output
-
-`prf_visit` lets a visitor interpret a raw block without placing Bloom, ORE,
-or other policy in this crate.
-
-```rust
-use vitaminc_prf::{HmacSha256Prf, PrfValue, PrfVisitor, PrfVisitorError};
-use vitaminc_protected::Protected;
-
-struct BloomPositions {
-    count: usize,
-    modulus: i16,
-}
-
-impl<P> PrfVisitor<[u8; 32], P> for BloomPositions {
-    type Value = Vec<i16>;
-
-    fn visit_block(self, block: [u8; 32]) -> Result<Self::Value, PrfVisitorError> {
-        Ok(block
-            .chunks_exact(2)
-            .take(self.count)
-            .map(|chunk| {
-                i16::from_le_bytes([chunk[0], chunk[1]]).rem_euclid(self.modulus)
-            })
-            .collect())
-    }
-}
-
-# async fn example() -> Result<(), Box<dyn std::error::Error>> {
-let prf = HmacSha256Prf::new(Protected::new(vec![7; 32]));
-let positions = "alice@example.com"
-    .prf_visit(
-        prf,
-        BloomPositions {
-            count: 4,
-            modulus: 2048,
-        },
-    )
-    .await?;
-
-assert_eq!(positions.len(), 4);
-# Ok(())
-# }
-```
 
 ## Implementing `PrfValue` for a struct
 
@@ -88,10 +28,9 @@ batch.
 
 ```rust
 use vitaminc_prf::{
-    BlockVisitor, HmacSha256Prf, IntoPrfContext, MapAccess, MapPrf, Prf,
+    BlockVisitor, IntoPrfContext, MapAccess, MapPrf, Prf,
     PrfValue, PrfVisitor, PrfVisitorError, SeqAccess,
 };
-use vitaminc_protected::Protected;
 
 struct User {
     email: String,
@@ -118,12 +57,6 @@ impl PrfValue for User {
     }
 }
 
-#[derive(Debug, PartialEq, Eq)]
-struct UserTerms {
-    email: [u8; 32],
-    aliases: Vec<[u8; 32]>,
-}
-
 struct BlockListVisitor;
 
 impl<P> PrfVisitor<[u8; 32], P> for BlockListVisitor {
@@ -132,6 +65,12 @@ impl<P> PrfVisitor<[u8; 32], P> for BlockListVisitor {
     fn visit_seq(self, seq: SeqAccess<[u8; 32], P>) -> Result<Self::Value, PrfVisitorError> {
         seq.map(|node| node.visit(BlockVisitor)).collect()
     }
+}
+
+#[derive(Debug, PartialEq, Eq)]
+struct UserTerms {
+    email: [u8; 32],
+    aliases: Vec<[u8; 32]>,
 }
 
 struct UserTermsVisitor;
@@ -156,18 +95,8 @@ impl<P> PrfVisitor<[u8; 32], P> for UserTermsVisitor {
         })
     }
 }
-
-# async fn example() -> Result<(), Box<dyn std::error::Error>> {
-let user = User {
-    email: "alice@example.com".into(),
-    aliases: vec!["alice".into(), "a.smith".into()],
-};
-let prf = HmacSha256Prf::new(Protected::new(vec![7; 32]));
-let terms = user
-    .prf_visit_with_context(prf, "tenant/acme/users/v1", UserTermsVisitor)
-    .await?;
-
-assert_eq!(terms.aliases.len(), 2);
-# Ok(())
-# }
 ```
+
+Executing the derivation requires a backend; with `vitaminc-hmac` in
+scope the value above resolves through
+`user.prf_visit_with_context(prf, "tenant/acme/users/v1", UserTermsVisitor).await`.
