+++ packages/prf/src/lib.rs
@@ -4,15 +4,18 @@
 mod context;
 mod encoding;
 mod error;
-mod hmac_sha256;
 mod impls;
+mod ready;
 mod traits;
 mod visitor;
+
+#[cfg(test)]
+mod test_backend;
 
 pub use context::{IntoPrfContext, PrfContext};
 pub use encoding::PrfEncoding;
 pub use error::{PrfBuildError, PrfError, PrfVisitorError};
-pub use hmac_sha256::{HmacSha256Prf, ReadyPrf};
 pub use impls::Passthrough;
+pub use ready::ReadyPrf;
 pub use traits::{MapPrf, Prf, PrfValue, SeqPrf};
 pub use visitor::{BlockVisitor, MapAccess, PrfVisitor, ResolvedPrf, ResolvedVisitor, SeqAccess};
