+++ release-plz.toml
@@ -43,6 +43,10 @@ name = "vitaminc-permutation"
 version_group = "vitaminc"
 
 [[package]]
+name = "vitaminc-prf"
+version_group = "vitaminc"
+
+[[package]]
 name = "vitaminc-protected"
 version_group = "vitaminc"
